+++ ArduCopter/AP_Arming.cpp
