+++ ArduCopter/AP_Arming.h
